20170612 wangxingwang 添加获取单个实例的方法

+++ svr/svr-ehr-basic/src/main/java/com.yihu.ehr/tj/dao/XTjDataSaveRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 /**
  * Created by Administrator on 2017/6/9.
  */
-public interface XTjDataSaveRepository extends PagingAndSortingRepository<TjDataSave, Integer> {
+public interface XTjDataSaveRepository extends PagingAndSortingRepository<TjDataSave, Long> {
 }
