+++ svr/svr-ehr-basic/src/main/java/com.yihu.ehr/tj/dao/XTjDataSourceRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 /**
  * Created by Administrator on 2017/6/9.
  */
-public interface XTjDataSourceRepository extends PagingAndSortingRepository<TjDataSource, Integer> {
+public interface XTjDataSourceRepository extends PagingAndSortingRepository<TjDataSource, Long> {
 }
